Move SkipLoggingHook to pkg/util/log
The apierrors-specific test are still in pkg/api/apierrors
