+++ pkg/images/deps/provider.go
@@ -9,7 +9,6 @@ import (
 
 	runtimeresource "github.com/authgear/authgear-server"
 
-	"github.com/authgear/authgear-server/pkg/api/apierrors"
 	imagesconfig "github.com/authgear/authgear-server/pkg/images/config"
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db"
@@ -46,7 +45,6 @@ func NewRootProvider(
 
 	loggerFactory := log.NewFactory(
 		logLevel,
-		apierrors.SkipLoggingHook{},
 		log.NewDefaultMaskLogHook(),
 		sentry.NewLogHookFromHub(sentryHub),
 	)
@@ -76,7 +74,6 @@ func NewRootProvider(
 func (p *RootProvider) NewAppProvider(ctx context.Context, appCtx *config.AppContext) *AppProvider {
 	cfg := appCtx.Config
 	loggerFactory := p.LoggerFactory.ReplaceHooks(
-		apierrors.SkipLoggingHook{},
 		log.NewDefaultMaskLogHook(),
 		config.NewSecretMaskLogHook(cfg.SecretConfig),
 		sentry.NewLogHookFromContext(ctx),
