+++ pkg/lib/deps/providers.go
@@ -7,7 +7,6 @@ import (
 	getsentry "github.com/getsentry/sentry-go"
 
 	"github.com/authgear/authgear-server"
-	"github.com/authgear/authgear-server/pkg/api/apierrors"
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/config/configsource"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db"
@@ -57,7 +56,6 @@ func NewRootProvider(
 
 	loggerFactory := log.NewFactory(
 		logLevel,
-		apierrors.SkipLoggingHook{},
 		log.NewDefaultMaskLogHook(),
 		sentry.NewLogHookFromHub(sentryHub),
 	)
@@ -93,7 +91,6 @@ func NewRootProvider(
 func (p *RootProvider) NewAppProvider(ctx context.Context, appCtx *config.AppContext) *AppProvider {
 	cfg := appCtx.Config
 	loggerFactory := p.LoggerFactory.ReplaceHooks(
-		apierrors.SkipLoggingHook{},
 		log.NewDefaultMaskLogHook(),
 		config.NewSecretMaskLogHook(cfg.SecretConfig),
 		// NewAppProvider is used in 2 places.
@@ -265,7 +262,6 @@ func NewBackgroundProvider(
 
 	loggerFactory := log.NewFactory(
 		logLevel,
-		apierrors.SkipLoggingHook{},
 		log.NewDefaultMaskLogHook(),
 		sentry.NewLogHookFromHub(sentryHub),
 	)
