+++ pkg/util/log/factory.go
@@ -22,6 +22,7 @@ func NewFactory(level Level, hooks ...logrus.Hook) *Factory {
 	logger.Out = ioutil.Discard
 	logger.Hooks.Add(&StackHook{})
 	logger.Hooks.Add(&ContextCauseHook{})
+	logger.Hooks.Add(&SkipLoggingHook{})
 	logger.Hooks.Add(NewWriterHook(os.Stderr))
 
 	for _, hook := range hooks {
