+++ pkg/util/log/writer_hook.go
@@ -45,8 +45,8 @@ func (h *WriterHook) Levels() []logrus.Level {
 }
 
 func (h *WriterHook) Fire(entry *logrus.Entry) error {
-	// The writer hook is supposed to be placed before SkipLoggingHook.
-	// So it does not skip logging any entry.
+	// WriterHook intentionally do not respect IsLoggingSkipped()
+
 	bytes, err := h.Formatter.Format(entry)
 	if err != nil {
 		return err
